change float to double

+++ include/asfproject/signal.h
@@ -13,9 +13,9 @@ namespace asf {
 class Wave;
 
 constexpr int DEF_FRAMERATE = 11025;
-constexpr float DEF_FREQ = 440.0;
-constexpr float DEF_AMP = 1.0;
-constexpr float DEF_PERIOD = 0.1F;
+constexpr double DEF_FREQ = 440.0;
+constexpr double DEF_AMP = 1.0;
+constexpr double DEF_PERIOD = 0.1;
 
 class Signal// NOLINT
 {
@@ -23,30 +23,30 @@ public:
   virtual ~Signal() = default;
 
   [[nodiscard]] virtual std::unique_ptr<Signal> clone() const = 0;
-  [[nodiscard]] Wave makeWave(float duration = 1.0, float start = 0, int framerate = DEF_FRAMERATE) const;
+  [[nodiscard]] Wave makeWave(double duration = 1.0, double start = 0, int framerate = DEF_FRAMERATE) const;
   void plot(int framerate = DEF_FRAMERATE) const;
-  [[nodiscard]] virtual float period() const;
-  virtual nc::NdArray<float> evaluate(const nc::NdArray<float> &ts) const = 0;// NOLINT
+  [[nodiscard]] virtual double period() const;
+  virtual nc::NdArray<double> evaluate(const nc::NdArray<double> &ts) const = 0;// NOLINT
 };
 
 class Sinusoid : public Signal//NOLINT
 {
 private:
-  float m_freq;
-  float m_amp;
-  float m_offset;
-  std::function<float(float)> m_func;
+  double m_freq;
+  double m_amp;
+  double m_offset;
+  std::function<double(double)> m_func;
 
 public:
-  explicit Sinusoid(float freq = DEF_FREQ,
-    float amp = DEF_AMP,
-    float offset = 0,
-    std::function<float(float)> func = static_cast<float (*)(float)>(std::sin));
+  explicit Sinusoid(double freq = DEF_FREQ,
+    double amp = DEF_AMP,
+    double offset = 0,
+    std::function<double(double)> func = static_cast<double (*)(double)>(std::sin));
   Sinusoid(const Sinusoid &other);
 
   [[nodiscard]] std::unique_ptr<Signal> clone() const override;
-  nc::NdArray<float> evaluate(const nc::NdArray<float> &ts) const override;//NOLINT
-  [[nodiscard]] float period() const override;
+  nc::NdArray<double> evaluate(const nc::NdArray<double> &ts) const override;//NOLINT
+  [[nodiscard]] double period() const override;
 };
 
 class SumSignal : public Signal//NOLINT
@@ -60,14 +60,14 @@ public:
   SumSignal(const SumSignal &other);
 
   [[nodiscard]] std::unique_ptr<Signal> clone() const override;
-  nc::NdArray<float> evaluate(const nc::NdArray<float> &ts) const override;//NOLINT 
-  [[nodiscard]] float period() const override;
+  nc::NdArray<double> evaluate(const nc::NdArray<double> &ts) const override;//NOLINT 
+  [[nodiscard]] double period() const override;
 };
 
 // Standalone functions
 
-std::unique_ptr<Sinusoid> cosSignal(float freq = DEF_FREQ, float amp = DEF_AMP, float offset = 0);
-std::unique_ptr<Sinusoid> sinSignal(float freq = DEF_FREQ, float amp = DEF_AMP, float offset = 0);
+std::unique_ptr<Sinusoid> cosSignal(double freq = DEF_FREQ, double amp = DEF_AMP, double offset = 0);
+std::unique_ptr<Sinusoid> sinSignal(double freq = DEF_FREQ, double amp = DEF_AMP, double offset = 0);
 std::unique_ptr<Signal> operator+(const Signal &sig1, const Signal &sig2);
 
 }// namespace asf
